Chart Display: Separate packages viewer.plot und viewer.plot.jfreechart.
Package net.imagej.viewer.plot.jfreechart contains two converter plugins:
 - Converter<XYPlot, JFreeChart>
 - Converter<CategoryPlot, JFreeChart>

Package net.imagej.viewer.plot contains a display for AbstractPlot using the two converters.
