+++ src/main/java/net/imagej/ui/swing/viewer/plot/DefaultPlotDisplay.java
@@ -8,13 +8,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -29,17 +29,24 @@
  * #L%
  */
 
-package net.imagej.ui.swing.viewer.plot.jfreechart;
+package net.imagej.ui.swing.viewer.plot;
 
 import net.imagej.plot.AbstractPlot;
-import net.imagej.plot.XYPlot;
+import net.imagej.ui.swing.viewer.plot.PlotDisplay;
+import org.scijava.display.AbstractDisplay;
 import org.scijava.display.Display;
+import org.scijava.plugin.Plugin;
 
 /**
- * Interface for {@link JfcPlotGenerator} {@link Display}s.
- * 
+ * Default display for {@link JfcPlotGenerator}s.
+ *
  * @author Curtis Rueden
  */
-public interface PlotDisplay extends Display<AbstractPlot> {
-	// This interface intentionally left blank.
+@Plugin(type = Display.class)
+public class DefaultPlotDisplay extends AbstractDisplay<AbstractPlot> implements
+		PlotDisplay
+{
+	public DefaultPlotDisplay() {
+		super(AbstractPlot.class);
+	}
 }
